requirements.txt
openai-whisper
gradio
edge-tts
jiwer
soundfile
numpy
pandas
pydub
